fix issue with import cycle
